Base timing system working

+++ ThePlatformer/System.h
@@ -14,7 +14,7 @@ namespace GameSystems
 			Success,
 			Error
 		};
-		virtual int Update(float dt, std::list<GameObjects::BaseGameObject*>&) = 0;
+		virtual int Update(double dt, std::list<GameObjects::BaseGameObject*>&) = 0;
 		virtual void Init(std::list<GameObjects::BaseGameObject*>&) = 0;
 		virtual void SendMessage() = 0;
 		virtual ~System() {}
